requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
httpx==0.25.2

# AWS dependencies
boto3==1.34.0
aws-cdk-lib==2.108.0
constructs==10.3.0

# Weather and geocoding
requests==2.31.0
geopy==2.4.0

# AI/ML dependencies
transformers==4.35.0
torch==2.1.0
spacy==3.7.2
nltk==3.8.1

# Natural language processing
spacy==3.7.2
nltk==3.8.1

# Date/time handling
python-dateutil==2.8.2
pytz==2023.3

# Environment and configuration
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.3
black==23.11.0
flake8==6.1.0 